fix: prevent null pointer when retrying events
When the first retry of an event occurs, the properties or headers object might not have been created yet. Yet we still use it in the logs. The msg object get default values in case these properties do not exist, but only in the errorhandler-function which is called from the function logging the retry-count. Moving the logic to add defaults to the errorhandler-function instead.